// backend/src/types/db.ts
import { ObjectId } from 'mongodb';

export type UserDocument = {
  _id: ObjectId;
  name: string;
  expectedOTJHours: number;
  actualOTJHours: number;
  totalOTJHours: number;
  lastOTJActivity: Date;
  createdAt: Date;
};

export type JournalEntryDocument = {
  _id?: ObjectId;
  userId: ObjectId;
  title: string;
  category: string;
  description: Description;
  timeSheets?: TimeSheet[];
  tagIds?: TagId[];
  createdAt: Date;
  updatedAt: Date;
};

export type Description = {
  intend: string;
  implementation: string;
  impact: string;
};

export type TimeSheet = {
  date: Date;
  startTime: Date;
  endTime: Date;
  duration: number;
};

export type TagId = string;

export type TagDocument = {
  _id: ObjectId;
  userId: ObjectId;
  tagDescription: string;
  tagColour: string;
};